Add menu.py
Implemented a terminal menu to execute the code in different modes, offering utility options and the ability to selectively run specific parts of the methodology. This section relied heavily on Al as it was completely new material for me. But I really wanted to implement this for better user experience